+++ server/src/models/ApiContract.ts
@@ -1,6 +1,5 @@
-import { PrismaClient } from '@prisma/client';
+import { prisma } from '../utils/prisma';
 
-const prisma = new PrismaClient();
 
 export interface ApiContract {
   id: number;